Add caveat about font-face format
Right now the example uses a `.woff` font with a format of `"woff"`. This can be misleading to a dev who doesn't know about font format keywords. I (and others https://stackoverflow.com/questions/76366169/custom-local-fonts-not-loading-in-astrojs-project) have wasted time here because fonts like `.otf` have a format of `"opentype"` instead of `"otf"`.